Include more information in the version subcommand

+++ psyopsOS/progfiguration/progfiguration_build.py
@@ -7,12 +7,12 @@ We support building for the following circumstances:
   This way the package can be installed locally, and told to build itself and install itself remotely on managed nodes.
 """
 
+import datetime
 import os
 import string
 import subprocess
-import time
+import textwrap
 import zipapp
-from typing import List
 
 
 PKG_ROOT = os.path.dirname(__file__)
@@ -22,26 +22,38 @@ APKBUILD_TEMPLATE = os.path.join(PKG_ROOT, "APKBUILD.template")
 APKBUILD_FILE = os.path.join(PKG_ROOT, "APKBUILD")
 
 
-def get_epoch_build_version():
-    # version = datetime.datetime.now().strftime("%Y%m%d.%H%M%S.0")
-    epochsecs = int(time.time())
+def get_epoch_build_version(dt: datetime.datetime):
+    """Return a version string based on the given datetime"""
+    epochsecs = int(dt.timestamp())
+    # version = dt.strftime("%Y%m%d.%H%M%S.0")
     version = f"1.0.{epochsecs}"
     return version
 
 
-def set_build_version(version: str):
+def set_build_version(dt: datetime.datetime):
     """Set the build version in the APKBUILD template and in the package itself.
 
     These files should not be added to git, take care.
     """
+    version = get_epoch_build_version(dt)
+
     with open(APKBUILD_TEMPLATE) as fp:
         apkbuild_template = string.Template(fp.read())
     apkbuild_contents = apkbuild_template.substitute(version=version)
-
-    with open(PKG_VERSION_FILE, "w") as fp:
-        fp.write(f"VERSION = '{version}'\n")
     with open(APKBUILD_FILE, "w") as fp:
         fp.write(apkbuild_contents)
+
+    package_version_template = string.Template(
+        textwrap.dedent(
+            """\
+            VERSION = '$version'
+            DATE = '$date'
+            """
+        )
+    )
+    package_version_contents = package_version_template.substitute(version=version, date=dt)
+    with open(PKG_VERSION_FILE, "w") as fp:
+        fp.write(package_version_contents)
 
 
 def build_alpine(apk_index_path: str, abuild_repodest: str = "psyopsOS", clean: bool = False):
@@ -51,10 +63,10 @@ def build_alpine(apk_index_path: str, abuild_repodest: str = "psyopsOS", clean: 
     abuild_repodest:    Set REPODEST as the repository location for created packages
     """
 
-    version = get_epoch_build_version()
+    buildtime = datetime.datetime.now()
 
     try:
-        set_build_version(version)
+        set_build_version(buildtime)
 
         if clean:
             print(f"Cleaning build in progfiguration directory...")
@@ -65,19 +77,21 @@ def build_alpine(apk_index_path: str, abuild_repodest: str = "psyopsOS", clean: 
         subprocess.run(abuild, cwd=PKG_ROOT)
     finally:
         failed = False
-        try:
-            os.unlink(PKG_VERSION_FILE)
-        except:
-            failed = True
-        try:
-            os.unlink(APKBUILD_FILE)
-        except:
-            failed = True
+        if os.path.exists(PKG_VERSION_FILE):
+            try:
+                os.unlink(PKG_VERSION_FILE)
+            except:
+                failed = True
+        if os.path.exists(APKBUILD_FILE):
+            try:
+                os.unlink(APKBUILD_FILE)
+            except:
+                failed = True
         if failed:
             raise Exception(
                 f"When trying to remove version file and/or ABKBUILD, got an exception. Manually remove these two files, if they exist: {PKG_VERSION_FILE}, {APKBUILD_FILE}"
             )
-    return version
+    return get_epoch_build_version(buildtime)
 
 
 def build_zipapp(outfile: str):
